using BigDecimal; included lombok and mapstruct

+++ src/main/java/ge/sibraine/cryptoinvestment/model/Crypto.java
@@ -2,14 +2,16 @@ package ge.sibraine.cryptoinvestment.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class Crypto {
 
     private String name;
-    private Double oldest;
-    private Double newest;
-    private Double min;
-    private Double max;
+    private BigDecimal oldest;
+    private BigDecimal newest;
+    private BigDecimal min;
+    private BigDecimal max;
     private Double normalizedRange;
 
 }
